storage: separate in-memory object reader responsibilities
