+++ packages/core/src/storage/objects/in-memory/index.ts
@@ -1,0 +1,2 @@
+export type { InMemoryObjectStorageSnapshot } from "./storage"
+export { getInMemoryObjectMaterializerAdapter, InMemoryObjectStorage } from "./storage"
